+++ src/Service/ArticleService.php
@@ -2,8 +2,59 @@
 
 namespace App\Service;
 
-class ArticleContentProvider
+use App\Entity\Article;
+use Doctrine\ORM\EntityManagerInterface;
+use App\Service\FileUploader;
+
+class ArticleService
 {
+    /**
+     * @var EntityManagerInterface
+     */
+    private $em;
+
+    /**
+     * @var FileUploader
+     */
+    private $articleFileUploader;
+
+    /**
+     * @param EntityManagerInterface $em
+     * @param FileUploader $articleFileUploader
+     */
+    public function __construct(EntityManagerInterface $em, FileUploader $articleFileUploader)
+    {
+        $this->em = $em;
+        $this->articleFileUploader = $articleFileUploader;
+    }
+
+    /**
+     * @param Article $article
+     * @param $image
+     * @param $plural
+     * @param $genitive
+     * @param $keywords
+     * @param $sizeFrom
+     * @param $sizeTo
+     * @param $theme
+     * @param $user
+     * @return void
+     */
+    public function create(Article $article, $image, $plural, $genitive, $keywords, $sizeFrom, $sizeTo, $theme, $user): void
+    {
+        $body = $this->generate_text($plural, $genitive, $keywords, $sizeFrom, $sizeTo, $theme);
+
+        $article
+            ->setDescription('Статья о ' . $keywords)
+            ->setBody($body)
+            ->setAuthor($user)
+            ->setPublishedAt(new \DateTime())
+            ->setImageFilename($this->articleFileUploader->uploadFile($image, $article->getImageFilename()));
+
+        $this->em->persist($article);
+        $this->em->flush();
+    }
+
     /**
      * @param $plural
      * @param $genitive
